docs: Merge cllct package doc to goq
